+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the Wazuh App for Splunk project will be documented in th
 - Outdated documentation links have been updated [#1290](https://github.com/wazuh/wazuh-splunk/pull/1290)
 
 ### Changed
+- Added the status `Pending` to the Agents sections. [#1292](https://github.com/wazuh/wazuh-splunk/pull/1292)
 - Added a disabled state to the `Apply changes` button on the Agents group editor when no changes on the group are made. [#1276](https://github.com/wazuh/wazuh-splunk/pull/1276)
 
 ## Wazuh v4.3.4 - Splunk Enterprise v8.1.[1-10], v8.2.x - Revision 4307